requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2 